chore: use 64-bit timestamps in RTMP provider stats

+++ src/projects/providers/rtmp/rtmp_stream.h
@@ -183,10 +183,10 @@ namespace pvd
 
 		uint32_t _key_frame_interval = 0;
 		uint32_t _previous_key_frame_timestamp = 0;
-		uint32_t _last_video_timestamp = 0;
-		uint32_t _last_audio_timestamp = 0;
-		uint32_t _previous_last_video_timestamp = 0;
-		uint32_t _previous_last_audio_timestamp = 0;
+		int64_t _last_video_timestamp = 0;
+		int64_t _last_audio_timestamp = 0;
+		int64_t _previous_last_video_timestamp = 0;
+		int64_t _previous_last_audio_timestamp = 0;
 		uint32_t _video_frame_count = 0;
 		uint32_t _audio_frame_count = 0;
